workspace reload
can now sorta load stored workspace via Gig -> LOAD (TEMP).  only loads setlist. the call to clear the old setlist is currently in the load method, because for some unknown reason if i run it elsewhere, the object gets restored between then and the call to create the new one, and the new one ends up second on the list and inaccessible for now. i think somewhere in there a callback is keeping the old object alive, investigating that now so i can move these calls to the correct place.