rollback preferences lift
preferences lift method was causing color wheel to potentially persist between runs. to resolve, instead of explicitly assigning the window to menu, i will have it assign its self.lift method as a callback in settings.windows, which gets triggered by any attempts to open a preferences window. when the window is closed, it will remove this callback in its quit_window method