x scroll & carriage return
implemented side to side scrolling with 'z' and '/' keys in prompt mode. useful in an emergency if you discover a line runs off the screen mid-show. currently using chunk scroll method but will probably switch to the behavior that "shift_scroll" uses where it advances while held, and is tied to the scroll speed setting.

also implemented "carriage return" which moves view all the way to the left and down a line. hit enter/return to advance a line and reset x_view.

a goal of these features is to allow arbitrary use of "ticker tape" style teleprompting mixed with the usual v scroll. though i've never seen it in practice, ticker tape prompting could be extremely useful in situations where only small screens are available.

eventually would like to implement an automatic mode which will scan across long lines when it reaches them, but manual is sufficient for now.

a nice side effect of tkinters h-scroll implementation is that you can not scroll beyond the last character on a visible line, so it is more difficult to go completely off the rails with this feature.