extracted gui
disentangled the tkinter gui from the MainApplication object, so now it lives in its own world. this will make it much easier when it is time to switch gui packages.

this of course required a lot of changes all over the place, so welcome to an unknown number of new bugs!

tried to test everything, everything that was working before this migration still seems to work. will take a day soon to actually set up unit tests.