pp style prompt toggle
added ctrl+k shortcut to immediately jump between fullscreen talent prompt and edit mode, which is essentially the ctrl+k behavior of the popular app presentation prompter.

in refactoring the edit toggle there is now a refresh issue on the edit/prompt toolbar - it doesn't always refresh until you click on it. fixing next before moving on to other features...