improved edit toggle
improved edit toggle calls, there was a double trigger in the menubar due to calling a function and linking the menu option to the variable directly.

haven't traced editbar refresh issue but noticed it only happens when there is no song loaded to editor.

discovered another bug with the ctrl+k method, which is, after a few rapid toggles the windowed mode of the talent window remains the size of the screen just with borders. i think this traces back to the way i capture window size before resize, and is probably a race condition. will investigate this along with the other issue before moving on