periodic editor refresh when scrolling
used a generator to only update editor y view every several frames. this proves slightly smoother on talent than 1:1 refresh but still not great. will move this option into preferences menu but leaving it on for now.

tried pushing talent y view into editor to see if it helped but it didn't...

continuing my investigation to see if there is an avoidable bottleneck somewhere...