building presentation prompter interface
want to make promptools work as a sidecar to presentation prompter for my dayjob. would like to be able to export setlists directly to PP format, which uses apple .plist to store bookmark / scroll speed / etc metadata. so i'm now building an interface for that.

currently mostly able to dump to plist file but there is this bytes object i haven't figured out how to decode / encode...